Update RestaurantView to have a PriceFilter component
